replaced list<int> in inputsUnfunden and replaced it with int[]

+++ Accounting/UTXO/Block.cs
@@ -13,9 +13,13 @@ namespace BToken.Accounting
     public UInt256 HeaderHash;
     public NetworkHeader Header;
     public List<TX> TXs;
-    public List<int[]> TXHashes;
+    public List<byte[]> TXHashes;
 
-    public Block(NetworkHeader header, UInt256 headerHash, List<TX> tXs, List<int[]> tXHashes)
+    public Block(
+      NetworkHeader header,
+      UInt256 headerHash, 
+      List<TX> tXs, 
+      List<byte[]> tXHashes)
     {
       Header = header;
       HeaderHash = headerHash;
